+++ src/main/java/gift/common/util/KakaoUtil.java
@@ -3,13 +3,11 @@ package gift.common.util;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.controller.oauth.KakaoProperties;
-import gift.controller.order.Link;
-import gift.controller.order.OrderRequest;
-import gift.controller.order.TextTemplate;
+import gift.controller.order.dto.Link;
+import gift.controller.order.dto.TextTemplate;
 import java.net.URI;
 import java.util.Map;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
