+++ src/main/java/gift/controller/order/OrderController.java
@@ -2,6 +2,8 @@ package gift.controller.order;
 
 import gift.common.annotation.LoginUser;
 import gift.common.auth.LoginInfo;
+import gift.controller.order.dto.OrderRequest;
+import gift.controller.order.dto.OrderResponse;
 import gift.service.OrderService;
 import java.net.URI;
 import org.springframework.http.ResponseEntity;
