+++ src/main/java/gift/controller/order/dto/Link.java
@@ -1,4 +1,4 @@
-package gift.controller.order;
+package gift.controller.order.dto;
 
 public record Link(String web_url) {
