+++ src/main/java/gift/controller/order/dto/OrderResponse.java
@@ -1,4 +1,4 @@
-package gift.controller.order;
+package gift.controller.order.dto;
 
 import gift.model.Order;
 import java.time.LocalDateTime;
