+++ src/main/java/gift/controller/order/dto/TextTemplate.java
@@ -1,4 +1,6 @@
-package gift.controller.order;
+package gift.controller.order.dto;
+
+import gift.controller.order.dto.Link;
 
 public record TextTemplate(
     String object_type,
