+++ src/main/java/gift/service/OrderService.java
@@ -4,8 +4,8 @@ import gift.common.auth.LoginInfo;
 import gift.common.exception.ErrorCode;
 import gift.common.exception.OptionException;
 import gift.common.util.KakaoUtil;
-import gift.controller.order.OrderRequest;
-import gift.controller.order.OrderResponse;
+import gift.controller.order.dto.OrderRequest;
+import gift.controller.order.dto.OrderResponse;
 import gift.model.Option;
 import gift.model.Order;
 import gift.repository.OptionRepository;
